+++ hibernate/src/main/java/by/andersen/training/hib/services/interfaces/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface UserService {
 
-    public User findById(int id);
+    public User findById(Long id);
 
     public void save(User user);
